# src/docs/swagger.yaml
definitions:
  model.InstanceVariable:
    properties:
      name:
        description: 变量名
        type: string
      type:
        description: 变量类型 1=int 2=string 3=bool 4=float64
        type: integer
      value:
        description: 变量值
        type: object
    type: object
  request.BatchSyncRoleUsersRequest:
    properties:
      roleUsersList:
        items:
          $ref: '#/definitions/request.SyncRoleUsersRequest'
        type: array
    type: object
  request.DenyInstanceRequest:
    properties:
      processInstanceId:
        description: 流程实例的id
        type: integer
      remarks:
        description: 备注
        type: string
    type: object
  request.HandleInstancesRequest:
    properties:
      edgeId:
        description: 走的流程的id
        type: string
      processInstanceId:
        description: 流程实例的id
        type: integer
      remarks:
        description: 备注
        type: string
      variables:
        description: 变量
        items:
          $ref: '#/definitions/model.InstanceVariable'
        type: array
    type: object
  request.ProcessDefinitionRequest:
    properties:
      classifyId:
        description: 分类ID
        type: integer
      formId:
        description: 对应的表单的id(仅对外部系统做一个标记)
        type: integer
      id:
        type: integer
      name:
        description: 流程名称
        type: string
      notice:
        description: 绑定通知
        type: string
      remarks:
        description: 流程备注
        type: string
      structure:
        description: 流程结构
        type: string
      task:
        description: 任务ID, array, 可执行多个任务，可以当成通知任务，每个节点都会去执行
        type: string
    type: object
  request.ProcessInstanceRequest:
    properties:
      processDefinitionId:
        description: 流程ID
        type: integer
      title:
        description: 流程实例标题
        type: string
      variables:
        description: 变量
        items:
          $ref: '#/definitions/model.InstanceVariable'
        type: array
    type: object
  request.SyncRoleUsersRequest:
    properties:
      roleId:
        type: integer
      userIds:
        items:
          type: integer
        type: array
    type: object
  response.HttpResponse:
    properties:
      data:
        type: object
      message:
        type: object
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
  /api/process-definitions:
    get:
      consumes:
      - application/json
      parameters:
      - description: 关键词
        in: query
        name: keyword
        type: string
      - description: 取的条数
        in: query
        name: limit
        type: integer
      - description: 跳过的条数
        in: query
        name: offset
        type: integer
      - description: asc或者是desc
        in: query
        name: order
        type: string
      - description: 排序键的名字，在各查询实现中默认值与可用值都不同
        in: query
        name: sort
        type: string
      - description: 类别 1=我创建的  2=所有
        in: query
        name: type
        type: integer
      - description: WF-TENANT-CODE
        in: header
        name: WF-TENANT-CODE
        required: true
        type: string
      - description: WF-CURRENT-USER
        in: header
        name: WF-CURRENT-USER
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.HttpResponse'
      summary: 获取流程定义列表
      tags:
      - process-definitions
    post:
      consumes:
      - application/json
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ProcessDefinitionRequest'
      - description: WF-TENANT-CODE
        in: header
        name: WF-TENANT-CODE
        required: true
        type: string
      - description: WF-CURRENT-USER
        in: header
        name: WF-CURRENT-USER
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.HttpResponse'
      summary: 创建流程模板
      tags:
      - process-definitions
    put:
      consumes:
      - application/json
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ProcessDefinitionRequest'
      - description: WF-TENANT-CODE
        in: header
        name: WF-TENANT-CODE
        required: true
        type: string
      - description: WF-CURRENT-USER
        in: header
        name: WF-CURRENT-USER
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.HttpResponse'
      summary: 更新流程模板
      tags:
      - process-definitions
  /api/process-definitions/{id}:
    delete:
      parameters:
      - description: request
        in: path
        name: id
        required: true
        type: string
      - description: WF-TENANT-CODE
        in: header
        name: WF-TENANT-CODE
        required: true
        type: string
      - description: WF-CURRENT-USER
        in: header
        name: WF-CURRENT-USER
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.HttpResponse'
      summary: 删除流程模板
      tags:
      - process-definitions
    get:
      parameters:
      - description: request
        in: path
        name: id
        required: true
        type: string
      - description: WF-TENANT-CODE
        in: header
        name: WF-TENANT-CODE
        required: true
        type: string
      - description: WF-CURRENT-USER
        in: header
        name: WF-CURRENT-USER
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.HttpResponse'
      summary: 获取流程模板详情
      tags:
      - process-definitions
  /api/process-instances:
    get:
      consumes:
      - application/json
      parameters:
      - description: 关键词
        in: query
        name: keyword
        type: string
      - description: 取的条数
        in: query
        name: limit
        type: integer
      - description: 跳过的条数
        in: query
        name: offset
        type: integer
      - description: asc或者是desc
        in: query
        name: order
        type: string
      - description: 排序键的名字，在各查询实现中默认值与可用值都不同
        in: query
        name: sort
        type: string
      - description: 类别 1=我的待办 2=我创建的 3=和我相关的 4=所有
        in: query
        name: type
        type: integer
      - description: WF-TENANT-CODE
        in: header
        name: WF-TENANT-CODE
        required: true
        type: string
      - description: WF-CURRENT-USER
        in: header
        name: WF-CURRENT-USER
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.HttpResponse'
      summary: 获取流程实例列表
      tags:
      - process-instances
    post:
      consumes:
      - application/json
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ProcessInstanceRequest'
      - description: WF-TENANT-CODE
        in: header
        name: WF-TENANT-CODE
        required: true
        type: string
      - description: WF-CURRENT-USER
        in: header
        name: WF-CURRENT-USER
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.HttpResponse'
      summary: 创建新的流程实例
      tags:
      - process-instances
  /api/process-instances/_deny:
    post:
      consumes:
      - application/json
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.DenyInstanceRequest'
      - description: WF-TENANT-CODE
        in: header
        name: WF-TENANT-CODE
        required: true
        type: string
      - description: WF-CURRENT-USER
        in: header
        name: WF-CURRENT-USER
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.HttpResponse'
      summary: 否决流程流程
      tags:
      - process-instances
  /api/process-instances/_handle:
    post:
      consumes:
      - application/json
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.HandleInstancesRequest'
      - description: WF-TENANT-CODE
        in: header
        name: WF-TENANT-CODE
        required: true
        type: string
      - description: WF-CURRENT-USER
        in: header
        name: WF-CURRENT-USER
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.HttpResponse'
      summary: 处理/审批一个流程
      tags:
      - process-instances
  /api/process-instances/{id}:
    get:
      parameters:
      - description: request
        in: path
        name: id
        required: true
        type: integer
      - description: request
        in: query
        name: includeProcessTrain
        type: boolean
      - description: WF-TENANT-CODE
        in: header
        name: WF-TENANT-CODE
        required: true
        type: string
      - description: WF-CURRENT-USER
        in: header
        name: WF-CURRENT-USER
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.HttpResponse'
      summary: 获取一个流程实例
      tags:
      - process-instances
  /api/process-instances/{id}/train-nodes:
    get:
      parameters:
      - description: request
        in: path
        name: id
        required: true
        type: integer
      - description: WF-TENANT-CODE
        in: header
        name: WF-TENANT-CODE
        required: true
        type: string
      - description: WF-CURRENT-USER
        in: header
        name: WF-CURRENT-USER
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.HttpResponse'
      summary: 获取流程链路
      tags:
      - process-instances
  /api/role-users:
    post:
      consumes:
      - application/json
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.SyncRoleUsersRequest'
      - description: WF-TENANT-CODE
        in: header
        name: WF-TENANT-CODE
        required: true
        type: string
      - description: WF-CURRENT-USER
        in: header
        name: WF-CURRENT-USER
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.HttpResponse'
      summary: 同步(创建或更新)角色用户映射关系
      tags:
      - role-users
  /api/role-users/_batch:
    post:
      consumes:
      - application/json
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.BatchSyncRoleUsersRequest'
      - description: WF-TENANT-CODE
        in: header
        name: WF-TENANT-CODE
        required: true
        type: string
      - description: WF-CURRENT-USER
        in: header
        name: WF-CURRENT-USER
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.HttpResponse'
      summary: 批量同步(创建或更新)角色用户映射关系
      tags:
      - role-users
swagger: "2.0"